Fixed checkout error handling
