+++ src/SprykerEco/Zed/Unzer/Business/ApiAdapter/Validator/UnzerApiAdapterResponseValidatorInterface.php
@@ -7,7 +7,6 @@
 
 namespace SprykerEco\Zed\Unzer\Business\ApiAdapter\Validator;
 
-use Generated\Shared\Transfer\CheckoutResponseTransfer;
 use Generated\Shared\Transfer\UnzerApiResponseTransfer;
 
 interface UnzerApiAdapterResponseValidatorInterface
@@ -20,15 +19,4 @@ interface UnzerApiAdapterResponseValidatorInterface
      * @return void
      */
     public function assertSuccessResponse(UnzerApiResponseTransfer $unzerApiResponseTransfer): void;
-
-    /**
-     * @param \Generated\Shared\Transfer\UnzerApiResponseTransfer $unzerApiResponseTransfer
-     * @param \Generated\Shared\Transfer\CheckoutResponseTransfer $checkoutResponseTransfer
-     *
-     * @return bool
-     */
-    public function isSuccessfulUnzerApiResponse(
-        UnzerApiResponseTransfer $unzerApiResponseTransfer,
-        CheckoutResponseTransfer $checkoutResponseTransfer
-    ): bool;
 }
